--- src/2023/TalkDetail/MeetingDetail.Type.ts
@@ -1,15 +1,0 @@
-import { SessionSpeaker } from "../Talks/Talk.types";
-
-export interface IMeeting {
-  urlName?: string;
-  title: string;
-  description: string;
-  videoUrl?: string;
-  slidesURL?: string;
-  videoTags?: string[];
-  speakers: SessionSpeaker[];
-  level?: string;
-  type?: string;
-  language?: string;
-  track?: string;
-}
