+++ src/views/MeetingDetail/MeetingDetail.tsx
@@ -5,7 +5,6 @@ import {
 } from "../../constants/BreakPoints";
 import {Color} from "../../styles/colors";
 import React, {FC, Suspense, useEffect} from "react";
-import {IMeeting} from "./MeetingDetail.Type";
 import LessThanIconWhite from "../../assets/images/LessThanIconWhite.svg";
 import LessThanIcon from "../../assets/images/LessThanBlueIcon.svg";
 import MoreThanIcon from "../../assets/images/MoreThanBlueIcon.svg";
@@ -37,7 +36,7 @@ import conferenceData from "../../data/2024.json";
 import {Tag} from "../../components/Tag/Tag";
 import styled from "styled-components";
 import {AddToCalendarButton} from "add-to-calendar-button-react";
-import {ISpeaker} from "../../types/speakers";
+import {IMeetingDetailProps, MyType} from "../../types/sessions";
 
 const getVideoHeight = (windowWidth: number) => {
   let videoHeight;
@@ -104,24 +103,6 @@ export const StyledVoteTalkLink = styled.a`
   color: ${Color.BLACK_BLUE};
   font-size: 0.8rem;
 `;
-
-interface IMeetingDetailProps {
-  meeting: IMeeting;
-  speakers?: ISpeaker[];
-}
-
-type MyType = {
-  urlName?: string;
-  videoUrl?: string;
-  level?: string;
-  videoTags?: string[];
-  speakers?: ISpeaker[];
-  description: string;
-  language?: string;
-  title: string;
-  type?: string;
-  track?: string;
-};
 
 const MeetingDetail: FC<React.PropsWithChildren<IMeetingDetailProps>> = ({
   meeting,
